requirements.txt
streamlit
google-generativeai
jinja2
python-docx
pdfkit
pyppeteer
fpdf